feat: 注釈の線幅と保存時外枠エフェクトを0.1px単位で調整可能に
- 線幅/外枠幅のスピンボックスを QDoubleSpinBox 化（0.1刻み）
- 線幅・外枠幅を float 化し、描画にアンチエイリアスを有効化
  （整数丸めを防ぎ 0.1px の差を実際に表現）
- 外枠は QRectF ベースの描画に変更、旧 int 設定との互換維持

+++ app/annotations.py
@@ -8,7 +8,7 @@ from PySide6.QtGui import QColor
 class RectAnnotation:
     rect: QRect
     color: QColor
-    line_width: int
+    line_width: float
 
     def copy(self) -> "RectAnnotation":
         return RectAnnotation(
